core/chains/evm/gas: remove OVM 1.0 gas estimator
